+++ CHANGES.rst
@@ -1,11 +1,12 @@
 CHANGES
 =======
 
-tip (unreleased)
-----------------
+master (unreleased)
+-------------------
 
 * Template tags now support Django's STATICFILES_STORAGE setting.
   Thanks Ivan Ven Osdel for report and fix.
+* Compatibility with Django 1.8 and 1.9.
 
 2.2.2 (2014.09.08)
 ------------------